Apply resource filtering by zim namespace
