+++ Model/WebViewConfig/WebViewConfiguration.swift
@@ -1,0 +1,46 @@
+// This file is part of Kiwix for iOS & macOS.
+//
+// Kiwix is free software; you can redistribute it and/or modify it
+// under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 3 of the License, or
+// any later version.
+//
+// Kiwix is distributed in the hope that it will be useful, but
+// WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+// General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Kiwix; If not, see https://www.gnu.org/licenses/.
+
+import Foundation
+import WebKit
+
+final class WebViewConfiguration: WKWebViewConfiguration {
+    override init() {
+        super.init()
+        setURLSchemeHandler(KiwixURLSchemeHandler(), forURLScheme: KiwixURLSchemeHandler.ZIMScheme)
+        #if os(macOS)
+        preferences.isElementFullscreenEnabled = true
+        #else
+        allowsInlineMediaPlayback = true
+        mediaTypesRequiringUserActionForPlayback = []
+        #endif
+        userContentController = {
+            let controller = WKUserContentController()
+            if let ruleList = WebContentBlocker.ruleList {
+                controller.add(ruleList)
+            }
+            if let url = Bundle.main.url(forResource: "injection", withExtension: "js"),
+               let javascript = try? String(contentsOf: url) {
+                let script = WKUserScript(source: javascript, injectionTime: .atDocumentEnd, forMainFrameOnly: false)
+                controller.addUserScript(script)
+            }
+            return controller
+        }()
+    }
+
+    required init?(coder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+}
